+++ src/Http/Requests/Request.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rubix\Server\Http\Requests;
+
+use GuzzleHttp\Psr7\Request as GuzzleRequest;
+
+class Request extends GuzzleRequest
+{
+    //
+}